Se ha añadido la clase BackTrackingMeshFilter.
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/branches/PFCSaraNew@505 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/Algorithms/BasicFilter/ImageCos.cpp
@@ -78,7 +78,7 @@ void ImageCos::TemplateProcess( int threadid)
                         extent.GetMin(0)*incx;
     for(x=extent.GetMin(0);x<=extent.GetMax(0); x++)
     {
-      *in_data1 = cos(*in_data1);
+      *in_data1 = cos((double)*in_data1);
       in_data1++;
     }
   }
